Change format to test incremental backup.  The "backup" configuration option is still there but is usually ignored now: since backups require logging, the default is to test if logging is configured.  (We could just remove the configuration option, but I thought I'd leave it in case it was useful some day.)
Add a new configuration "timer=#", which specifies the length of the run
in minutes.  This is a potentially significant change because it changes
how often format discards session/cursor handles and checkpoints, so it
may change the characteristics of the run.
Specifically:
	Previously, each worker thread did a first checkpoint somewhere
	between rand(1, N-operations), now it's between rand(100, 10K)
	operations, and threads did a subsequent checkpoint every
	(rand(1, N-operations) / 5), now it's every rand(5K, 20K) operations.

	Previously, each worker thread open/closed its session/cursor
	handles between rand(1, N-operations) if single-threaded, and
	between rand(100, 5K) operations if not single-threaded, now
	it's always between rand(100, 5K), regardless.

Change the "master" thread to figure out when the worker threads should
quit rather than the worker threads figuring it out on their own, it's
possibly a bit slower, but it's a simpler solution.

Align the worker thread reporting structure on 64B, avoid cache line
overlap.

Change the tracking mechanism, once a number gets to the 10M range, report
it in millions, otherwise we run off the edge of the "standard" window.

Allow explicit configuration of backups or compaction when single threaded.