Flush the turtle file to disk even if checkpoint-sync isn't configured, otherwise we could end up with a zero-length file after the rename.
I'm not positive we need to sync the directory if checkpoint-sync isn't
configured, my reading of POSIX 1003.1 is we're guaranteed one file or
the other after a rename. But, I'm not confident in that reading, and
there's some discussion in the POSIX groups that indicates confusion as
to what exactly is required from and/or provided by different filesystems.