Re-commit the change to avoid a handle deadlock.
There is a bug I can't see and missed in the review.

Original comment:
Avoid a deadlock between the sweep server and a checkpoint.

The sweep server was holding a handle exclusive and waiting on the handle list lock (in order to remove the handle from the list), a checkpoint was holding the handle list lock waiting for a read lock on the handle.  Found by examining a hang during automated testing.