WT-1967
When updating pages in the service of the lookaside table, the page/tree
has to remain clean, because if the handle is a checkpoint handle, it
has no power to write the tree. Split page-modification into two parts,
the marking of the page/tree as dirty, and actual page modification.
The cursor code does both, as does re-instantiating pages in the split
code, the lookaside table code only updates the page.

There are a couple of issues:

First, we mark the page/tree dirty before making the first modification
to the page, if the first modification fails for some reason, the
page/tree would remain dirty, the page's first-dirty transaction might
be slightly earlier than otherwise.

Incidentally, this fixes a minor bug where the memory costs from the
first set of modifications were added as "clean byte" costs in the cache
because the page hadn't yet been marked dirty. There are reasons we did
that in older versions of the system, but that was before the cache
memory tracking was redone.  I removed the comment about "incrementing
the in-memory footprint after releasing the mutex", that's no longer
relevant because we no longer lock in many (most?) paths.

Second, the actual page-modification code can no longer depend on the
page having a WT_PAGE_MODIFY structure: there are a few places in the
serialization code that now have to test before using the structure (for
example, when acquiring page locks). For that specific example, we could
move the page lock byte into the WT_PAGE structure (there's room), but
the WT_PAGE_MODIFY.obsolete_check_txn field is probably not going to
move, and the underlying update-serialization function uses it.

I cleaned up serial.i a bit to use the WT_ERR style coding conventions.