WT-2076: two changes the help with the slowdown.
I'm pushing this branch because it appears to fix the problems, but I
don't fully understand why these changes are necessary.

First, instead of checking for 2 6th level skiplist entries, test for 8
5th level entries. I'm seeing wtperf workloads with 30K entries in the
skiplist (level 0), and a level 6 with no entries (where the expected
value is roughly 7). Testing the master branch (currently cc4ad7f), it
doesn't appear to happen, I don't see empty level 6's there.

Second, don't call __wt_page_can_split() to test for in-memory splits
unless we're actually interested in doing an in-memory split. This is
another change that's confusing, the code is unchanged since before the
slowdown. My guess is we now call __wt_page_can_evict in a way where
returning "can evict" because the in-memory split check was successful
makes a difference, a test if eviction would succeed was less of a
problem before.