SERVER-23622 Allow eviction to consider pages from files being checkpointed.
Otherwise eviction will never queue clean pages from files while they
are being checkpointed. This causes problems in some MongoDB workloads.
In particular workloads where the oplog is generating most of the
work in a checkpoint, and is the major source of read pressure
(i.e: clean pages) in the cache.