WT-2603: Track the hash bucket ID for hash queues, not the name hash value.
Change WiredTiger hash bucket implementations to track the bucket ID
instead of the name hash.

We don't resize the hash queues so there's never a need to recompute the
bucket ID.

Set a flag on the metadata data-handle during the open instead of doing
a comparison on the name all the time.