WT-2672 handle system calls that don't set errno
Define system call success as a 0 return, and split error handling
into two parts: if the call returns -1, use errno, otherwise expect
the failing return to be an error value.

Replace calls to remove with unlink, so we know errno will be set.
Do the best we can with rename, there's no easy workaround.