WT-2711 statistics log paths
Change the statistics_log path configuration to be a directory name,
remove the file name component.

This potentially requires an upgrade step for applications that were
specifying a non-standard file name component for statistics log file
names, it's not backward compatible.

Add a discussion of absolute path names to the security section of the
documentation.