WT-2665 Limit allocator fragmentation from the WiredTiger cache (#2799)
* Add a statistic for the amount of cache used for page images vs other objects

* Reset the "maximum page size" statistic each checkpoint

* Add a separate queue for urgent eviction, replaces existing WOULD_BLOCK eviction mode

Until now, the eviction server has had to walk the cache to find pages that
would otherwise block application threads.  This change allows threads to put
those pages directly on a queue that is checked before ordinary LRU eviction.

* Don't queue pages for urgent eviction when eviction is disabled in a tree.

Decouple evict_queue_lock and the individual queue's evict_lock in
__evict_get_ref.  Don't block the eviction server if there are many
application threads waiting on a queue.

* Take care with the urgent queue.

Entries in the urgent queue cannot be skipped or they will not be considered again.