Minor cleanup of checkpoint configuration.
Do all configuration steps before checking to see if the checkpoint
server is going to run (don't assume the application does all of its
checkpoint configuration in a single call).

Don't reset the log information if we're not going to run a checkpoint
server.

Don't test for in-memory configuration incompatibility until after
deciding if we're going to run the checkpoint server (not a bug fix, but
it's cleaner not to repeat the test). Reword the error: checkpoints are
the problem here, not the in-memory configuration.