Revert back to stopping/restarting the statistics_log server any time there's a reconfigure; otherwise the configuration routine is going to have to free memory from the configuration structure as it parses the configuration information (not too hard), but also we'd have to deal with races if the running server is looking at the configuration information that's changing (too hard).
Reconfiguration processing is fundamentally different if we ever stop and
then restart a server as part of reconfiguration. The problem is the stop
step will discard configuration information that's been copied into the
connection structure, so we have to review all of the configuration strings
to re-instantiate that information in the connection structure.  The log
server function can continue to skip configuration strings that aren't
supposed to appear during reconfiguration, because there's no stop/restart
step, the connection information remains correct, the log server can ignore
strings that shouldn't appear during reconfiguration. The statistics-log
server cannot do that, it must retrieve the values configured as part of
the original statistics-log setup (from the saved database configuration
values), and re-instantiate them in the connection structure.  In summary,
revert back to always checking statistics_log.path in the statistics_log
configuration routine, relying on configuration declarations to ensure
statistics_log.path is never set during reconfiguration.