WT-2764 Replace write leaves phase of checkpoints.
Instead, get multi-threaded writes by dialling down the eviction dirty
trigger.

Also included is a fix to operations that could checkpoint the metadata
before a checkpoint is committed (e.g., creating a new table when
logging is disabled).  Such operations should include the checkpoint
transaction ID in their snapshot so that the in-progress metadata
updates aren't flushed before the checkpoint is durable.