WT-2767 in test/suite/run.py, add -s N option to run an individual scenario (#2878)
* Added -s N option to run.py to restrict run to a scenario with a
given number or name. When printing the name of the test, if there
is a scenario, print as 'fulltestname -s scenario_number' so it can be
easily copied for another command run.  Created a 'testname.txt' file
in the test directory with the name of the test (and '-s NNN', if appropriate).
Fixed the ordering of the symbolic name of a scenario (e.g. 'file.recno.hex'),
it was reversed before.

* Only allow -s when a test is specified (not for a complete run).

(cherry picked from commit 27614f9928e9c69cb1f4c0535e479b29af11f23c)