WT-2737 Scrub dirty pages rather than evicting them
Link/include test/testutils with wtperf so we can use the test utility
macros and functions in wtperf.

Add the -R option to wtperf to remove and re-create the run directory
so it's not nececessary to do it outside of wtperf.

Minor changes to wtperf variables, sort some options, make some things
boolean.

Merge the stub memory allocation functions used by the test code and
wtperf.