Add a new reconciliation flag WT_EVICT_DISK_IMAGE; it causes reconciliation to save disk images it creates. Change the eviction of multi-block pages to optionally re-instantiate pages in memory instead of evicting them.
Remove the WT_EVICT_IN_MEMORY flag: it's identical to WT_CONN_IN_MEMORY,
use WT_CONN_IN_MEMORY instead.

When instantiating pages in-memory, set the read-generation to the same
value as the read-generation of the original page (unless it's set to
WT_READGEN_OLDEST, in which case leave the page's read-generation unset.