Revert fad8ea1, we want to copy the disk image even if we find a match in a previous write.
The bug was in setting the boundary structure from the multi-block
reconciliation structure. First, the boundary structure should already
have been set (and setting it to the multi-block information leaked
allocated memory). Second, the multi-block information should be NULL,
which caused us to copy compressed disk images into the disk image
structure, which will fail when we re-instantiate the page.