Only assert if we are actually copying values.
The case we care about is a cursor pointing to data on a page without a
transaction ID pinned.  It is okay for cursors to be positioned without
a transaction ID pinned if it has a copy of the value.  This is the
difference between WT_CURSTD_VALUE_INT and WT_CURSTD_VALUE_EXT.