WT-2804 Allow non-transactional reads from checkpoints.
Improve two recent assertions, one from WT-2798 relating to writing
metadata updates to disk that are part of a running transaction, and
another from WT-2802 that checks that we don't try to copy values from a
cursor without a transaction pinned.  The latter doesn't apply to
cursors on checkpoints (including chunk cursors in an LSM tree).