WT-2842 split wtperf's configuration into per-database and per-run parts
Create a CONFIG_OPTS structure, which is a structure that holds all of
the configuration options, and change the CONFIG structures to reference
it. In other words, instead of each configuration option being a
CONFIG.field, it's a CONFIG.opts.field, and all of the CONFIG structures
reference the same CONFIG.opts structure.

Move config_copy(), config_free() and config_compress() from config.c to
wtperf.c, they're CONFIG structure functions, not CONFIG_OPT functions.

Move usage() from config.c to wtperf.c, it's not a generic config function.

Rename/Rework configuration functions to be consistently config_opt_XXX
functions, and to minimize use of the CONFIG structure.