The build_posix Make.base file (the source of the top-level Makefile), tries to build docs/index.html if that file is missing, by running dist/s_docs, and the s_docs script compiles the wtperf/doxy.c program to output a doxygen formatted list of wtperf configuration options.
The doxy.c program has become tricker to build in the current branch,
it needs WiredTiger-specific types. While it's arguably impossible for
Make.base to create docs/index.html anyway (there's no reason to believe
doxygen is installed during any particular compile), make it work again.

Replace the doxy.c program with a Python script so we don't have to
compile anything in the dist scripts. (We still require cpp, the C
preprocessor command, hopefully that's not a deal-breaker anywhere.)