WT-2923 heap-use-after-free on address in compaction
We can't look at the WT_REF address without some kind of lock as the
address can change underneath us. For example, if we take a copy of
WT_REF.addr.addr while it's referencing an on-page cell, and then the
page is evicted before we crack that address, we'll potentially read
free'd memory. Lock down the WT_REF before reading the address.