WT-2954 Try to evict pages that exceed memory_page_max even if the transaction ID isn't increasing
Workloads that insert large values checking only for the transaction ID
moving forward can lead to very large in-memory pages. Allow the page to exceed
the split size, but not the configured maximum.

Skipping such pages is a performance optimization - allow the optimization to
work as long as pages are between split size and maximum size.

Also don't require so many entries on a leaf page to trigger a split when
there are large values.