Allow variables to be used in ifdefs by defining them to the C preprocessor. Ignore a single ';' in the .run file, which may be an artifact of a macro that has been 'ifdef-ed' out.
Take advantage of the above to create a maintainable .run file that works
on systems without ftruncate. Use macros to avoid using so many #ifdefs.

When output differences are detected, report the output's line number (or
range of lines numbers we tried to match) to make it easier to diagnose
differences.