Change compaction to use database-wide checkpoints: rather than processing each file separately and checkpointing after each time we compact a file, do a compaction pass over all of the files then database-wide checkpoints.
Change the general checkpoint function to support a new boolean argument,
"waiting", if the caller wants to wait until the checkpoint lock is
available, or return immediately.

Add a "skip this handle" boolean to the data-handle structure so we can
to skip compacting file handles where we're not doing useful work.

Add a "checkpoint is running" boolean to the global transaction
structure so compaction can wait on an already-running checkpoint.

Change to a finer-granularity compaction timeout by checking every 100
pages we walk in the cache.

Replace the LSM compaction timeout code with a call to a common function.