With the change to call __wt_block_compact_start/end() from the upper-level session compaction code, we should have moved the compaction verbose statistics reset from __wt_block_compact_skip() to __wt_block_compact_start().
The skip function is called at the start of every compaction pass, and
the verbose statistics have to persist over all of the passes, not just
one. (The statistics could be per pass and then this change wouldn't be
necessary, but that's not how it's set up, and resetting in the skip
call both loses all previous values and ensures they're zero when we do
display them because the last pass is always skipped because no more
compaction will be useful, leaving the stats set to zero).