WT-3030 Fix a race between scans and splits reading the index hint.
After a split, child references in an internal page can end up in a
different location in the index array.  WiredTiger maintains a hint of
the expected offset, and searches if the expected reference is not
found at the location of the hint.

This change fixes a race where multiple threads could set the hint to
different values (if they are searching different versions of an
index).  This could lead to a search returning the wrong slot, because
we were writing the correct index into the hint, *then* reading from
the hint into the returned value.