Initial implementation of growing the hazard pointer array.
Deprecate the hazard_max configuration, default to an initial 250 slots
and grow from there.

Remove the check in eviction where we skip files if we have used all
available hazard pointers.

Remove the write-barrier (WT_PUBLISH) before incrementing the hazard_inuse
value. We haven't written anything so I can't see why a write-barrier
helps. It's been that way since the original commit of the code (de5376d,
September 25, 2012), and I'll call it out in the review for a specific check.

Note these changes are incomplete, if the hazard pointer array grows,
the original memory is simply leaked, it's never freed.