Coverity complains: 	CID 1368529:  Security best practices violations  (TOCTOU) 	Calling function "fopen" that uses "fname" after a check 	function. This can cause a time-of-check, time-of-use race 	condition.
We're doing:

	snprintf(buffer);
	stat(buffer);
	snprintf(buffer);
	fopen(buffer);

and I think Coverity is ignoring the second snprintf(), and is complaining
about a stat followed by an fopen some number of lines of code later.  It's
simple enough to give the two calls their own buffers, hopefully that will
keep Coverity quiet.

Use 1024 as the size of a path instead of 512, (that's the traditional
MAXPATHLEN value).

Use sizeof(home) in calls to testutil_work_dir_from_path() so we don't
accidentally diverge from the declared size.

Clean up an error call, there's no need for two error messages.