Coverity complains:
	CID 1368528:    (DEADCODE)
	Execution cannot reach the expression "","" inside this
	statement: "pos += (uint32_t)snprintf(c...".

Replace boolean variable with a "const char *" that's set to either an
empty string or a comma, removing the need for the test.

Use size_t as the size of an object in memory, not a uint32_t.

Don't declare variables in block scope.

Assignment operators are the lowest priority operator (well, except for
comma), don't bother declaring the order of evalution for an assignment
operator.

strlen() returns a size_t length, don't evaluate it as a boolean.