Utility command error handling cleanup:
Always print an error message if a WT_SESSION method fails, we don't
know if the WiredTiger library printed out a message or not.

Free memory allocated by util_uri() in some cases where we either didn't
have the necessary free call, or simply returned without freeing memory.

Try and be more consistent with error messages, use the leading call as
the first string, and any arguments to that call as the second.

Replace some of the places we're writing to stderr explicitly with the
utility error handlers.

Initialize the return variable from util_uri() in all cases.

Change error messages that referenced WT_SESSION.open to reference
WT_SESSION.open_cursor.