After 0b9e453, we no longer need to do any non-DIAGNOSTIC work after completing the split (previously, we had changes to make the newly created split pages evictable, but now they are initially given a generation number which will prevent their eviction until it's OK).
Rename __split_ref_step2() to be __split_verify_intl(), and change it
to verify all of the internal pages involved in the split.

Previously, we only verified the pages we had to read and update anyway.
Now we don't have to update any pages and we're only reading pages in
DIAGNOSTIC mode, verify all of them.

Don't release the hazard pointer explicitly, use the more standard
__wt_page_release() call (it should make no difference, it's just a bit
more consistent).

Rename __split_ref_step1() to be __split_ref_prepare(), there's no
longer a step #2.