Noticed search termination conditions when sampling the search page vs. walking it sequentially weren't the same. Changed that, which caused the test_compact02 test to fail.
There's an underlying bug in this code, if we return WT_NOTFOUND, we can
lose a hazard pointer on the page of the tree we unsucessfully searched.
Add a page-release in the case of returning not-found.

Simplify the termination conditions as well: there are bunch of ref
states I'd just as soon not worry about, for example, WT_REF_SPLIT.  I
think they'd all work, except split, but as I said, I'd just as soon not
worry about them. Limit our valid returns to in-memory if we're running
on behalf of eviction, disk and in-memory for everybody else.