Don't jump out of the loop if we find an invalid root page, return WT_NOTFOUND instead. Jumping out of the loop leads to potentially setting a returned WT_REF and returning 0, which aren't right for __wt_btcur_next_random().
For clarity, this change is not part of the bug fix: we should never
be in this code unless we're in eviction, but I don't want to return
success, either.

We weren't always overwriting the returned WT_REF; do so with a
NULL so no caller depends on it not being overwritten in some path.

Fix a comment.