WT-3203 bulk-load state changes can race
The bulk-load state change (where multiple threads of control turn off
the possibility of bulk-load in a tree), has always been able to race,
but it's potentially dangerous now that turning off bulk-load involves
calling `__wt_evict_file_exclusive_off()`. In the current tree, there's
a diagnostic-only test that might drop core because of this race.

Add a bulk-load spin lock to the WT_BTREE structure, use it to ensure we
don't race when updating the bulk-load state.