Change the WT_CURSOR.remove semantics to retain the cursor position if the call starts with the cursor positioned, otherwise don't maintain a cursor position.
I've moved the set/clear of WT_CURSTD_KEY_SET, WT_CURSTD_VALUE_SET from
the cursor level into the btree level, but only for the WT_CURSOR.remove
call -- the rest will follow in another change set.

While in the area, I noticed that WT_CURSOR.insert on a table followed
by an iteration failed to start at the beginning/end of the table, add
a fix for that.