Changing the btree cursor remove code to use the correct "is the cursor positioned" test (WT_CURSTD_KEY_INT), breaks some things:
First, the cursor level remove code does a copy of any WT_CURSTD_KEY_INT
key when it's called, use WT_CURSOR_CHECKKEY() in that code instead of
WT_CURSOR_NEEDKEY() so the right flag survives into the btree layer.

Second, the btree truncate code assumes calling the btree cursor remove
code will leave it with a positioned cursor; that's no longer the case
if truncate doesn't start with a positioned cursor, and because of how
the cursor layer currently works, we're not starting with a positioned
cursor. In general, it's a bad idea to depend on truncate's callers to
pass in a positioned cursor, change the btree truncate code to call
cursor search instead of cursor remove, that way truncate is guaranteed
a positioned cursor for a small additional cost.