WT-3206 Fix a race allocating split generations.
We use split generations to detect when readers may be looking at
structures that are replaced by a split.  For correctness, we should
only increment the global split generation *after* a split becomes
public.  Only then can we safely check that no thread is still reading
with the old generation.

Previously, a split could increment the global split generation, then a
thread could start reading with the new split generation but see the old
index structure.

This issue was introduced by WT-3088, where we wanted a way to ensure
that newly-allocated pages don't split until it is safe.  That is solved
here by having the split code pin a split generation in the ordinary way
(without allocating a new one) for the duration that splits of new
pages need to be prevented.