Only check for an empty tree the first time an LSM chunk is opened.
The goal here is to make sure that LSM primary chunks start empty.
Otherwise, we can't load into a skiplist in memory as required by LSM.
If an operation such as verify closes a btree in order to check the
on-disk state, the next time it is opened we have to check whether it is
empty.

It is safe to do this check without locking: what matters is that we
always do the `lsm_primary` check before any update operation that would
turn off `btree->bulk_load_ok`.