Illustrates bug in Swig generation of Java Cursor methods
This commit adds JUnit tests that illustrate a bug in the Swig
type map logic for Java. The existing type map passes all methods
with int return values through logic that throws a WiredTigerException
if the return value is non-zero. This follows the standard C
approach of returning non-zero where Java might throw instead.

However, there are two Cursor methods where the authors intended
to actually return the result of the method as an integer to Java.
This can be seen because they duplicated the logic from the
type map into the compare_wrap and equals_wrap methods.

This commit illustrates the bug. My next commit will fix it