The leaf-page value dictionary stores cell offsets in the disk image, which implies a dictionary reset any time we hit a boundary or grow the disk image buffer. Recent changes broke that, we weren't resetting the dictionary when the disk image buffer was resized.
Instead of clearing the dictionary on buffer resize, switch to using cell
offsets in the dictionary instead of cell pointers. It's unlikely to be a
big win for many workloads, but it might help some, and it's cleaner than
resetting the dictionary more often.

Add a verify of disk images we don't write: the I/O routines verify any image
we write, but we need to verify any image we create.