WT-3245 Avoid hangs on shutdown when a utility thread encounters an error
Panic if the log server threads exit unexpectedly.

Remove an async worker test for panic -- it doesn't make a difference if
an async worker thread exits in the case of a panic, but it doesn't gain
us anything either, might as well have the async threads look the same
as the other server threads.