Review comments:
Rename __wt_session_gen_{publish,clear} to __wt_session_gen_{enter,leave},
change to return a void for simplification.

Change the thread enter function to loop until it sees the published
generation is still current, to ensure there isn't a race when scanning
for the oldest generation.

Remove the version of the function that also incremented the resource
generation, we don't use it yet and it's easy to add if we ever need it.