WT-3292 review/cleanup full-barrier calls in WiredTiger
Rework logging's log_close_fh and log_close_lsn barrier code. There are two
fields, the close handle and the close LSN, and they both have to be set.
The __log_file_server() code was checking the close handle and then spinning
until the close LSN was set, but a simpler solution is to publish the close
LSN write before setting the close handle.

To be clear, the previous code was correct, it was the comment that was wrong.