SERVER-28166 Assess effects of pinning a lot of content in storage engine cache
Fix some problems with the LAS table.

The generalized generation code sets initial generations to 1, which
causes the checkpoint collision test in __rec_las_checkpoint_test() to
fail, because the generation doesn't match the WT_BTREE structure's
initial generation of 0. Fix by setting each new WT_BTREE structure's
generation to the current checkpoint generation.

Also, minor shuffling to put the WT_BTREE structure's checkpoint fields
together.

Reconciliation maintains a count during eviction of NOT skipped update
chains (that is, update chains that didn't contain an uncommitted
transaction), and uses it to set a return status if the lookaside table
might be useful after a failed eviction. (This is a bit confusing, the
variable tracking the count is named "update_skip_cnt".) However, the
count is only incremented if there were no skipped updates AND all of
the updates a globally visible. That's wrong, we can use the LAS table
if there are no skipped updates regardless of global visibility, and
checking if all of the updates are globally visible means we're much
less likely to select LAS.

Also, switch the variable to counting skipped updates instead of NOT
skipped updates, to match the name.

Reconciliation has a test in __rec_write_check_complete() of whether an
update-restore page reconciliation should fail because there aren't
enough updates being discarded that it's worth rewriting the page
in memory. That test incorrectly used the count of currently allocated
boundary structures in its calculations, not the count of boundary
structures used in a specific reconciliation. That bug led to eviction
repeatedly doing update-restore reconciliation and rewriting
pages in memory, to little or no gain, instead of switching to the LAS
table.

Further, the test was intended to be if 10% of the update chains won't
have to be re-instantiated when the page is rewritten. I think the test
was wrong, but regardless, implementing the test correctly doesn't help
with the test load I'm using. (My test is creating a snapshot and then
inserting records until the cache is full.) Re-instantiating the page
in memory stalls in that test case, which makes sense because rewriting
a page in memory isn't evicting anything, it's just shrinking the memory
being pinned down. Change the test: if we can discard a disk-page size
chunk of the original page, rewrite the page in memory.  Else, fallback
to the lookaside table if we think lookaside will work.  Else, rewrite
the page if 10% of the update chains won't have to be re-instantiated.